update (SMAPI 3.18.4, SDV 1.5.6)

+++ source/~daleao/Modular Overhaul/Modules/Professions/Events/GameLoop/DayEnding/ConservationismDayEndingEvent.cs
@@ -56,7 +56,7 @@ internal sealed class ConservationismDayEndingEvent : DayEndingEvent
             return;
         }
 
-        var trashCollectedThisSeason = player.Read<uint>(DataKeys.ConservationistTrashCollectedThisSeason);
+        var trashCollectedThisSeason = (int)player.Read<float>(DataKeys.ConservationistTrashCollectedThisSeason);
         player.Write(DataKeys.ConservationistTrashCollectedThisSeason, "0");
         if (trashCollectedThisSeason <= 0)
         {
